+++ src/main/java/org/apache/jackrabbit/filevault/maven/packaging/AbstractPackageMojo.java
@@ -103,7 +103,7 @@ public abstract class AbstractPackageMojo extends AbstractMojo {
     /**
      * Defines the content package type. this is either 'application', 'content', 'container' or 'mixed'.
      * If omitted, it is calculated automatically based on filter definitions. certain package types imply restrictions,
-     * for example, 'application' and 'content' packages are not allowed to contain sub packages or embedded bundles.<br/>
+     * for example, 'application' and 'content' packages are not allowed to contain sub packages or embedded bundles.<br>
      * Possible values:
      * <ul>
      *   <li>{@code application}: An application package consists purely of application content. It serializes
@@ -191,6 +191,15 @@ public abstract class AbstractPackageMojo extends AbstractMojo {
         Filters filters = new Filters();
         filters.load(getFilterFile());
         return filters;
+    }
+    
+    static File getFirstExistingDirectory(File[] directories) {
+        for (File dir: directories) {
+            if (dir.exists() && dir.isDirectory()) {
+                return dir;
+            }
+        }
+        return null;
     }
 
  }