+++ src/main/java/org/apache/jackrabbit/filevault/maven/packaging/Embedded.java
@@ -16,15 +16,6 @@
  */
 package org.apache.jackrabbit.filevault.maven.packaging;
 
-import java.util.ArrayList;
-import java.util.List;
-import java.util.Set;
-
-import org.apache.jackrabbit.filevault.maven.packaging.impl.StringFilterSet;
-import org.apache.maven.artifact.Artifact;
-import org.apache.maven.artifact.resolver.filter.ScopeArtifactFilter;
-import org.apache.maven.project.MavenProject;
-
 /**
  * The {@code Embedded} class represents an embedded artifact dependency
  * from the project descriptor. Such an embedding is declared in
@@ -35,35 +26,14 @@ import org.apache.maven.project.MavenProject;
  *     &lt;groupId&gt;artifact.groupId.pattern&lt;/groupId&gt;
  *     &lt;artifactId&gt;artifact.artifactId.pattern&lt;/artifactId&gt;
  *     &lt;scope&gt;compile&lt;/scope&gt;
- *     &lt;type&gt;jar&lt;/type&gt;
- *     &lt;classifier&gt;sources&lt;/classifier&gt;
+ *     &lt;type&gt;artifact.type.pattern&lt;/type&gt;
+ *     &lt;classifier&gt;artifact.classifier.pattern&lt;/classifier&gt;
  *     &lt;filter&gt;true&lt;/filter&gt;
  *     &lt;target&gt;/libs/sling/install&lt;/target&gt;
  * &lt;/embedded&gt;
  * </pre>
  */
-public class Embedded {
-
-    /**
-     * A group filter string, consisted of one or several comma separated patterns.
-     */
-    private final StringFilterSet groupId = new StringFilterSet();
-
-    /**
-     * A artifact filter string, consisted of one or several comma separated patterns.
-     */
-    private final StringFilterSet artifactId = new StringFilterSet();
-
-    private ScopeArtifactFilter scope;
-
-    private String type;
-
-    private String classifier;
-
-    /**
-     * If {@code true} a filter entry will be generated for all embedded artifacts.
-     */
-    private boolean filter;
+public class Embedded extends SimpleEmbedded{
 
     /**
      * Target location.
@@ -74,28 +44,6 @@ public class Embedded {
      * Name to use for the artifact in the destination
      */
     private String destFileName;
-
-    private boolean excludeTransitive;
-
-    public void setGroupId(String groupId) {
-        this.groupId.addEntries(groupId);
-    }
-
-    public void setArtifactId(String artifactId) {
-        this.artifactId.addEntries(artifactId);
-    }
-
-    public void setScope(String scope) {
-        this.scope = new ScopeArtifactFilter(scope);
-    }
-
-    public void setAddFilter(boolean filter) {
-        this.filter = filter;
-    }
-
-    public boolean isFilter() {
-        return filter;
-    }
 
     public String getDestFileName() {
         return destFileName;
@@ -118,50 +66,10 @@ public class Embedded {
         return target;
     }
 
-    public void setExcludeTransitive(boolean excludeTransitive) {
-        this.excludeTransitive = excludeTransitive;
-    }
-
-    public boolean isExcludeTransitive() {
-        return excludeTransitive;
-    }
-
-    public List<Artifact> getMatchingArtifacts(final MavenProject project) {
-
-        // get artifacts depending on whether we exclude transitives or not
-        final Set<Artifact> deps;
-        if (excludeTransitive) {
-            // only direct dependencies, transitives excluded
-            deps = project.getDependencyArtifacts();
-        } else {
-            // all dependencies, transitives included
-            deps = project.getArtifacts();
-        }
-
-        final List<Artifact> matches = new ArrayList<Artifact>();
-        for (Artifact artifact : deps) {
-            if (groupId.contains(artifact.getGroupId())
-                    && artifactId.contains(artifact.getArtifactId())
-                    && (scope == null || scope.include(artifact))
-                    && (type == null || type.equals(artifact.getType()))
-                    && (classifier == null || classifier.equals(artifact.getClassifier()))) {
-                matches.add(artifact);
-            }
-        }
-        return matches;
-    }
-
     public String toString() {
         final StringBuilder builder = new StringBuilder();
         builder.append("Embedded: ");
-        builder.append("groupId=").append(groupId).append(",");
-        builder.append("artifactId=").append(artifactId).append(",");
-
-        if (scope != null) {
-            builder.append("scope=").append(scope).append(",");
-        }
-        builder.append("filter=").append(filter);
-        builder.append(",excludeTransitive=").append(excludeTransitive);
+        super.toString(builder);
         if (target != null) {
             builder.append(",target=").append(target);
         }
@@ -170,4 +78,5 @@ public class Embedded {
         }
         return builder.toString();
     }
+
 }